Config subscripe for Render

+++ src/app/services/user.service.ts
@@ -1,7 +1,8 @@
-import {Injectable} from '@angular/core'
-import {HttpClient} from '@angular/common/http'
-import {Router} from '@angular/router'
+import { Injectable } from '@angular/core'
+import { HttpClient } from '@angular/common/http'
+import { Router } from '@angular/router'
 import * as API from '../services/config'
+import { Subject } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -10,38 +11,56 @@ export class UserService {
 
   currentUser: {}
 
-  constructor(private http: HttpClient, private _Router: Router){
+  subEventAddUser: Subject<any> = new Subject<any>() // Add User
+
+  subEventRejectUser: Subject<any> = new Subject<any>()// Destroy User
+
+  constructor(private http: HttpClient, private _Router: Router) {
+
+    this.subEventRejectUser.subscribe(value => {
+      if (value) {
+        localStorage.clear()
+        this.currentUser = {}
+        let url = this._Router.url // url current
+        // this._Router.navigateByUrl('/home', { skipLocationChange: true }).then(() =>
+        //   this._Router.navigate([url]))
+        // //console.log(url)
+        this._Router.navigateByUrl(url)
+      }
+    })
+
+    this.subEventAddUser.subscribe(value => {
+      let fullname = value['fullname']
+      let token = value['token']
+      localStorage.setItem('_fullname', fullname)
+      localStorage.setItem('_token', token)
+      this.currentUser = { fullname, token }
+    })
+
     let fullname = localStorage.getItem('_fullname')
     let token = localStorage.getItem('_token')
+
     this.currentUser = {
-      fullname,token
+      fullname, token
     }
   }
 
-  Login(user: FormData){
+  Login(user: FormData) {
     let urlAPI = `${API.urlAPI}/account/login`
-    return this.http.post(urlAPI,user)
+    return this.http.post(urlAPI, user)
   }
 
-  setUser(fullname,token){
-    localStorage.setItem('_fullname',fullname)
-    localStorage.setItem('_token',token)
-    this.currentUser={
-      fullname,token
-    }
+  setUser(fullname, token) {
+    this.subEventAddUser.next({ fullname, token }) // setUser
   }
 
-  Logout(){
-    localStorage.clear()
-    this.currentUser={}
-    let url = this._Router.url // url current
-    console.log(url)
-    this._Router.navigate([url])
+  Logout() {
+    this.subEventRejectUser.next(true) // Destroy User
   }
 
-  Register(user: FormData){
+  Register(user: FormData) {
     let urlAPI = `${API.urlAPI}/account/register`
-    return this.http.post(urlAPI,user)
+    return this.http.post(urlAPI, user)
   }
 
 }
